#207 Added realm validation on raw trail deletion (#208)
* #207 Added realm validation on raw trail deletion

* #207 Added error validation test

+++ backend/src/main/java/org/sc/controller/admin/AdminTrailRawController.java
@@ -3,6 +3,7 @@ package org.sc.controller.admin;
 import io.swagger.v3.oas.annotations.Operation;
 import org.sc.common.rest.response.TrailRawResponse;
 import org.sc.controller.response.TrailRawResponseHelper;
+import org.sc.data.validator.auth.AuthRealmValidator;
 import org.sc.manager.TrailRawManager;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.DeleteMapping;
@@ -11,6 +12,7 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
 import java.util.Collections;
+import java.util.Set;
 
 import static org.sc.controller.Constants.ONE;
 import static org.sc.controller.Constants.ZERO;
@@ -22,20 +24,27 @@ public class AdminTrailRawController {
 
     private final TrailRawManager trailRawManager;
     private final TrailRawResponseHelper trailRawResponseHelper;
+    private final AuthRealmValidator authRealmValidator;
 
     @Autowired
     public AdminTrailRawController(final TrailRawManager trailRawManager,
-                              final TrailRawResponseHelper trailRawResponseHelper) {
+                                   final TrailRawResponseHelper trailRawResponseHelper,
+                                   final AuthRealmValidator authRealmValidator) {
         this.trailRawManager = trailRawManager;
         this.trailRawResponseHelper = trailRawResponseHelper;
+        this.authRealmValidator = authRealmValidator;
     }
 
     @Operation(summary = "Delete a single raw trail")
     @DeleteMapping("/{id}")
     public TrailRawResponse deleteById(final @PathVariable String id) {
-        return trailRawResponseHelper
-                .constructResponse(Collections.emptySet(),
-                        trailRawManager.deleteById(id),
-                        ONE, ZERO, ONE);
+        final Set<String> errors = authRealmValidator.validate(id);
+        if(errors.isEmpty()) {
+            return trailRawResponseHelper
+                    .constructResponse(Collections.emptySet(),
+                            trailRawManager.deleteById(id),
+                            ONE, ZERO, ONE);
+        }
+        return trailRawResponseHelper.constructResponse(errors, Collections.emptyList(), ZERO, ZERO, ONE);
     }
 }
